factor out cuda code
